Big changes
Lade till map_object som superklass till interactable, character och
block. Den har en funktion poly(), som måste overrideas för att kunna
använda objekt av typen som önskas. Påbörjat kollisionshanterare för
breakable.

+++ SuperFoobar/Interactable-filer/Finish_line.h
@@ -4,6 +4,6 @@
 class Finish_line : protected Interactable
 {
 public:
-	Finish_line(float, float, float, float);
+	Finish_line(int, int, int, int);
 	~Finish_line() = default;
 };
