+++ SuperFoobar/Interactable-filer/Interactable.cpp
@@ -1,8 +1,10 @@
-#include "../Interactable-filer/Interactable.h"
+//#include "../Collision/Cluster.h"
+#include "Interactable.h"
+
 
 using namespace std;
 
-Interactable::Interactable(float x_pos, float y_pos, float height, float width)
+Interactable::Interactable(int x_pos, int y_pos, int height, int width)
 	: position{ new Cluster(x_pos, y_pos, height, width) }
 {}
 
@@ -16,12 +18,12 @@ void Interactable::interact()
 	dead = true;
 }
 
-float Interactable::get_width()
+int Interactable::get_width()
 {
 	return width;
 }
 
-float Interactable::get_length()
+int Interactable::get_length()
 {
 	return length;
 }
